# api/openapi/xflow-v1.yaml
openapi: 3.1.0
info:
  title: xflow Workflow Management API
  version: 1.0.0
  description: |
    User-facing HTTP contract for xflow. This spec mirrors the routes that are
    actually registered on the apiserver mux (docs/design/API-SPECIFICATION.md
    §7, §10). The runner protocol face — including POST /v1/executions
    (entry-seed) and the entire /v1/runners/* family — is intentionally absent:
    per spec §0.1 / §10 the OpenAPI contract describes only the user face.

    Every user-face JSON response is wrapped in the Envelope (§3). The only
    exceptions are the raw-byte streams (GET /v1/supplies/{name},
    GET|HEAD /v1/artifacts/{digest}) whose 2xx bodies are octet-streams (their
    failure bodies are still JSON Envelopes — spec §3.4) and the /healthz and
    /readyz probes, which load balancers parse directly and must not be
    enveloped (§7).

    Delivery semantics, stated once here instead of per endpoint: the handler
    boundary and the Runner Protocol are at-least-once — lease replay, node
    execution timeout and failover can each invoke the same handler twice — so
    business side effects are the host's responsibility, with
    `execution_id` + `node_name` as the recommended idempotency key. Those are
    the faces this document deliberately omits (spec §0.1 / §10); see
    docs/design/RELEASE-GATES.md §3 and §3.1. No endpoint below promises
    exactly-once delivery.
  contact:
    name: xflow maintainers
    url: https://github.com/xbcio/xflow
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/management/runner-pools/{id}:
    get:
      operationId: getRunnerPool
      summary: Get a runner pool
      description: Returns the pool when it is visible to the caller; hidden and unknown pools are both 404.
      tags: [Management]
      parameters:
        - $ref: "#/components/parameters/RunnerPoolId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Runner pool.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunnerPoolView"
        default:
          $ref: "#/components/responses/DefaultError"
    patch:
      operationId: updateRunnerPool
      summary: Update a runner pool
      description: Mutable settings may change, but allowed namespaces and node types may only be narrowed.
      tags: [Management]
      parameters:
        - $ref: "#/components/parameters/RunnerPoolId"
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerPoolUpdateRequest"
      responses:
        "200":
          description: Updated runner pool.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunnerPoolView"
        default:
          $ref: "#/components/responses/DefaultError"
    delete:
      operationId: deleteRunnerPool
      summary: Soft-delete a runner pool
      description: Soft-deletes a visible pool; subsequent reads and enrollments treat it as not found.
      tags: [Management]
      parameters:
        - $ref: "#/components/parameters/RunnerPoolId"
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Deletion acknowledged.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunnerPoolMutationResponse"
        default:
          $ref: "#/components/responses/DefaultError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-issued bearer token bound to a (subject, namespace, scopes) principal.
  parameters:
...
    RunnerPoolId:
      name: id
      in: path
      required: true
      description: Runner pool id.
      schema:
        type: string
...
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Optional client-supplied request id (charset [A-Za-z0-9._-], length
        1..128). When valid it is echoed verbatim in the response header for
        client-side retry correlation (spec §5.2). Never appears in the body.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
...
  headers:
    XRequestId:
      description: Echoed X-Request-Id (only when the inbound value was valid).
      schema:
        type: string
...
  responses:
    DefaultError:
      description: Error response — a failure Envelope with success=false and no data field.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      description: |
        The user-face response envelope (spec §3). On success success=true,
        code="200" (string), and data carries the payload. On failure
        success=false, data is omitted (not null) — except for
        workflow_param_invalid, which carries data.param_issues — and code
        carries a stable snake_case business identifier. trace_id is server-authoritative
        (W3C 32-hex).
      required: [success, code, message, trace_id]
      properties:
        success:
          type: boolean
          description: Strictly consistent with the HTTP status (2xx → true).
        code:
          type: string
          description: Machine-readable identifier. "200" on success; stable snake_case business code on failure.
        message:
          type: string
          description: Human-readable text. "" on success. Never carries credentials, node output, SQL, or stack traces (§3.5).
        data:
          description: Payload on success. Omitted on failure, except workflow_param_invalid (see ParamInvalidErrorEnvelope).
        trace_id:
          type: string
          description: Server-authoritative W3C trace id (32 hex).
          pattern: '^[0-9a-f]{32}$'
    ErrorEnvelope:
      type: object
      description: The failure shape of the Envelope — data is absent on failure.
      additionalProperties: false
      required: [success, code, message, trace_id]
      properties:
        success:
          type: boolean
          description: Always false on this shape.
          example: false
        code:
          type: string
          description: Stable snake_case business error code.
          example: workflow_not_found
        message:
          type: string
          description: Human-readable text. May change; clients must branch on code.
          example: workflow not found
        trace_id:
          type: string
          description: Server-authoritative W3C trace id (32 hex).
          pattern: '^[0-9a-f]{32}$'
...
    RunnerPoolUpdateRequest:
      type: object
      description: Every field is optional. Scope fields may only narrow their current values.
      properties:
        name:
          type: string
          minLength: 1
        allowed_namespaces:
          type: array
          items:
            type: string
        allowed_node_types:
          type: array
          items:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        max_instances:
          type: integer
          minimum: 0
        inherit_namespaces:
          type: boolean
        paused:
          type: boolean
    RunnerPoolView:
      type: object
      required: [id, name, owner_kind, owner_namespace, allowed_namespaces, allowed_node_types, labels, max_instances, inherit_namespaces, paused, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        owner_kind:
          type: string
          enum: [tenant, platform]
        owner_namespace:
          type: string
          description: Empty for platform-owned pools.
        allowed_namespaces:
          type: array
          items:
            type: string
        allowed_node_types:
          type: array
          items:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        max_instances:
          type: integer
          minimum: 0
        inherit_namespaces:
          type: boolean
        paused:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    RunnerPoolMutationResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [deleted, revoked]
...